# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GLOWNDRY Laundry Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all laundry services
      responses:
        "200":
          description: Services list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Bookings ──────────────────────────────────────────────────────────────────
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List all payments (admin)
      responses:
        "200":
          description: Payments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a payment (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ── Notifications ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Service:
      type: object
      required: [id, name, description, pricePerKg, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        pricePerKg:
          type: number
        createdAt:
          type: string
          format: date-time
    ServiceInput:
      type: object
      required: [name, description, pricePerKg]
      properties:
        name:
          type: string
        description:
          type: string
        pricePerKg:
          type: number
    ServiceUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        pricePerKg:
          type: number

    # ── Booking ──────────────────────────────────────────────────────────────────
...
    Payment:
      type: object
      required: [id, bookingId, amount, cashReceived, change, createdAt]
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        customerName:
          type: ["string", "null"]
        amount:
          type: number
        cashReceived:
          type: number
        change:
          type: number
        createdAt:
          type: string
          format: date-time
    PaymentInput:
      type: object
      required: [bookingId, amount, cashReceived]
      properties:
        bookingId:
          type: integer
        amount:
          type: number
        cashReceived:
          type: number

    # ── Notification ─────────────────────────────────────────────────────────────
...